pyproject.toml
[project]
name = "pdf-utils-web"
version = "0.2.0"
description = "PDF & Image processing web app with Gemini OCR and MCP server"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "Naveen Reddy", email = "naveenreddy61@gmail.com" }
]
dependencies = [
    "aiosqlite>=0.21.0",
    "fastmcp>=2.10.6",
    "google-cloud-storage>=3.9.0",
    "google-genai>=1.31.0",
    "httpx>=0.28.1",
    "litellm>=1.75.0",
    "pillow>=11.3.0",
    "playwright>=1.58.0",
    "PyMuPDF>=1.24.0",
    "pymupdf4llm>=0.0.27",
    "pypandoc>=1.15",
    "python-dotenv>=1.1.1",
    "python-fasthtml>=0.9.0",
    "pyyaml>=6.0.2",
    "reportlab>=4.4.3",
    "tiktoken>=0.9.0",
    "trafilatura>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/naveenreddy61/anki_nav_mcp_server"
Repository = "https://github.com/naveenreddy61/anki_nav_mcp_server.git"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
pdf-mcp-server = "pdf_mcp_server:main"
pdf-web-app = "web_app.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/pdf_mcp_server", "src/web_app", "src/pdf_utils"]

[tool.uv]
package = true
